Add files via upload
The notification.py file in the Flask-based backend is responsible for handling the functionality of sending notifications. These notifications can be in the form of emails (using services like SendGrid) or SMS messages (using services like Twilio). The purpose of this file is to encapsulate the logic for sending notifications, so other parts of the application can easily use it to alert users about certain events (such as receiving mail).