Add files via upload
The requirements.txt file is an essential component in Python projects, especially for Flask-based applications. It contains a list of all the dependencies (i.e., external libraries) that the project needs to run. Each dependency listed in this file can be installed using the pip package manager. By having this file, other developers can quickly set up the project environment by installing the exact versions of the libraries required for the project.